build_src/requirements.txt
Flask
prometheus_client
xrpl-py
python-dotenv